10.17.2024 Update
Revised the app.py file and removed routes and pages that were not being used anymore. I added market capitalization, edited the create account pages to give an error if the username or email was already in use, and I added a route to create an admin account from the admin market page. I also added the portfolio page that Seth worked on to my files.