10.22.24 Massive update
In this update users can sign in as admin or customer, and I also changed the navbar to have a user icon and to reduce the number of routes people were able to click on. Customers can now only go to the portfolio page and market page once they are signed in. Customers and admins cannot access pages unless they are signed in, and customers cannot access the admin pages on admins