10.30.2024 Updates
The stock current price is now fluctuating when the virtual internment is running, I also was able to get the market hours to restrict users from buying stocks based on the day of the week and the time of day. I also added the US holidays and set it to close the market when there is a US holiday.  Imported 3 new python libraries to achieve this.  

from apscheduler.schedulers.background import BackgroundScheduler
from datetime import datetime, time, date
import holidays