10.31.24 Updates
Created a new table and flask wtf form in the app.py file for the cash account. The users can now add money to their cash accounts and have it displayed on the portfolio page. There is also a confirmation message that is displayed when they submit button is clicked.