11.11.2024 Updates
Added CSS styling to the buying and selling stock pages. The pages now show the current price of the stock the user clicked on, the users available balance, and the total cost of the stock is calculated based on the volume the user selects. I also added custom error messages to the buy and sell pages for when the user is trying to buy stock but they don't have enough funds or if they are trying to sell more stock volume then they currently have. The custom messages provide a better user experience since it doesn't user flash messages and it won't all populate on the market page.